Allow github forks to run travis nosetests builds
The wiki user 'Pywikibot-test' only works on the 'wikimedia'
travis-ci account.  This prevents forks of the gibhub from
running tests on travis-ci with the original codebase.

nosetests support was added recently, and can be used to run
only tests which do not access the Wikimedia sites or other
Internet services.

Change-Id: I0f22caaf8b5cd314a196a5e3cc25ae830231ccb5